Refactor code
- Moved functions to modules - file handling
- Renamed function validate_tgt_dir to validate_out_dir
